Amend README to tie this application with ds-caselaw-public-ui
This application is intended to work with the Marklogic database from
ds-caselaw-public-ui. However the README instructions gave the impression that
a separate ML instance should be spun up for this app.

I think it's safe to assume that anyone running this application has already
built ds-caselaw-public-ui, so amend the README to make it explicit that the
ML instance this application is the same one as for ds-caselaw-public-ui

Also, remove the ML container from docker-compose, as running this container
conflicts with the one in ds-caselaw-public-ui