Remove links to "Terms of Use" page
Rollbar: https://rollbar.com/dxw/tna-caselaw-editor-ui/items/7/

The "Terms of Use" page does not have a template in this project (it exists in
the public UI project). We do not need a "Terms of Use" page for the Editor UI
as it is not part of the public website, so we can remove any references to it.