Explicitly append new elements to the XML tree when updating
This PR relies on https://github.com/nationalarchives/ds-caselaw-custom-api-client/pull/35
being merged and a new release of the API client

If a metadata element does not exist in the XML document (as opposed to existing
but being empty) we were previously returning `None` from the API client and
the document was therefore uneditable. Now, we will return an empty element if
the element doesn't exist in the document. The element (new or existing) can
then be added to the element tree and re-saved back into the document.