Edit the Judgment metadata on the API Client side
Bump the API Client to v4.5.1

Previously we were editing the Judgment XML in the editor using ETree, and
saving it back to the database via the client. This was causing namespaces
to be lost when deserialising the XML.

In the latest API Client version (4.5.1) we have enabled single-node editing
using XQuery, which preserves the namespaces in the XML document as a whole.
This should fix various issues we've seen with namespaces vanishing, and
data not appearing on the public and editor UIs.