If a PDF exists for a judgment, show the Download PDF button
We will eventually have PDFs in the public asset bucket for a judgment. If a
PDF exists, show a Download PDF button, so that the editor can view it.

Unlike a similar commit in the public UI - https://github.com/nationalarchives/ds-caselaw-public-ui/pull/232/
- we do not want to offer the generated PDF for editors. The generated PDF
is a stopgap for the public UI until we get the "real" PDFs in the S3 bucket.