Download the Judgment PDF from the PRIVATE_ASSET_BUCKET
In an earlier PR we added the ability to download the Judgment PDF from the
public asset bucket. This is incorrect - editors shoud download the PDF from the
*private* asset bucket, so they can check it before publication.

We have refactored this to work more in line with downloading the docx -
rather than getting the PDF from the bucket and returning it in an HTTPResponse,
we can generate a presigned URL and download the PDF in the same way we download
the docx. This simplifies the views and urls considerably.