Only treat a judgment as a total "failure" if it has no judgment XML
If a judgment has XML that starts with the `akomantoso` tag, treat it as a valid
judgment, even if it has a `failure` URI.

If the root of the judgment XML is `error` (i.e. there was no valid XML in the
ingested tarball) then treat it as a "red failure".

Otherwise, show the transformed HTML for the judgment and allow the judgment to
be edited etc.