Don't return the string "None" if any of the editable details are missing
If the Neutral Citation in a judgment was missing, the edit page was showing the
string "None" in the edit box. We want to show an empty string for a NoneType/
falsy type, not the string "None", so add a default to be returned if any of
these values are a NoneType.