Copy a judgment's properties alongside the judgment content
When we copy a judgment from one URI to anotherr, we should also copy its
properties (which live outside the content).

It woul have been neater to copy the properties XML in the API Client, but
we have a time pressure here and it is eaasier to use the pre-existing
`set_property` API client call to do this.