Correctly handle a NotFound error from Marklogic
In a previous commit, we were assuming that a Not Found response from Marklogic
would be returned as an empty judgment (""). However what is actually returned
is a MarklogicApiError.

We need to handle this response. if a MarklogicApiError is returned,
try to move the judgment to  new URI.

If a MarklogicApiError is *not* returned, we assume there is a judgment at that
URI, show an error to the editor informing them that there is already a judgment
at that URI, and do not attempt to move the "failure" judgment