Restore the Delete button for "failure" judgments
Trello: https://trello.com/c/ZfEJWIbN

In an earlier commit, we removed the "Delete" button for "failure" judgments
which had legitimate judgment content in them. We only showed the delete button
for judgments which were "total" failures (i.e. there was no judgment xml).

The thinking behind this was that once a judgment was edited and had a genuine
URI, the "failure" would be deleted behind the scenes.

However, it appears that some "failure" judgments are duplicates, and will not
be edited as there is already a judgment in Marklogic with that NCN/URI. These
judgments need to be deleted.

Therefore, restore the Delete button for any judgment with "failure" in the
URI.