Guard against deleted judgments appearing in search results
Something I have noticed on dev/staging but not production - deleted judgments
appearing in the search results. I am not sure if Marklogic caches search
results?

In the meantime this change guards against the side effects if this is the case.
If a judgment appears in search results, we try to build a SearchResult
model from it. If we cannot (because the judgment does not exist) a
MarklogicApiError is thrown, which we catch and return None.

If the list of SearchResult objects contains a None, filter it out before
the search results are rendered to the page.

Rollbar: https://rollbar.com/dxw/tna-caselaw-editor-ui/items/50/