Method to build a new AWS key from the new URI
When we copy assets from a "failure" location to a new URI-based location in
AWS S3, we need to change the keys for all assets, and the filename for the
docx assets. The ingester will name the docx after the "failure" URI, so it
needs to be renamed to a name in line with the URI (see ds-caselaw-ingester
for the logic).

This function will rename a docx in accordance to the URI, for example changing
`failures/TDR-2022-DNWR/failures_TDR-2022-DNWR.docx` to
`/ukpc/2023/120/ukpc_2023_120.docx` if the new URI is `/ukpc/2023/120`

Any non-docx assets, for example images, will keep their filename, e.g.
`failures/TDR-2022-DNWR/image1.jpg` will become `/ukpc/2023/120/image1.jpg`