Copy assets to new URI location when judgment is moved to a new URI
When a judgment is moved from a "failure" URI (e.g. `failures/TDR-2022-DNWR`) to
an NCN-derived URI (e.g. `ukpc/2023/120`) we want to move the assets (docx,
images, pdf etc) to an S3 bucket location that matches the URI.

Move the assets when the judgment is successfully moved in Marklogic.