Add last page to paginator
Duplicate of https://github.com/nationalarchives/ds-caselaw-public-ui/pull/284

If there were 25 judgments, only two pages would be shown on the paginator
due to an off-by-one error: list(range(1,5)) returns [1, 2, 3, 4]. We
now add one to the maximum value.

This has the side effect of making page 20 visible from page 10 when there
are a large number of pages. This feels more like the correct behaviour
to me.

The 2000-page pagination test has been increased to 2500 to test behaviour
when there are more judgements than the paginator wants to show; additional
tests now exist for 5 and 25 judgments to test smaller edge cases.