(Fix) When moving & retreiving items in S3, strip the leading slash from keys
While debugging the functionality to move assets around in S3 when a judgment is
moved, we found that the S3 moving & retrieval functions are very sensitive
to a leading slash on the uri. For example looking for a prefix of `/xxx/2022`
in S3 returns nothing but `xxx/2022` returns items.

Additionally, the docx/pdf URL generation will look for a filename with a leading
underscore if a leading slash is on the URI.

To mitigate this, strip the URI of a leading slash whenever we use it to
find items in S3.