Set internal URI in judgments when they are moved to a new URI
When a "failure" judgment is moved to a new URI in Marklogic, set the new
URI in the judgment XML. This is so that the URI can be used to build the src
URL for images inside the document.

v4.7.0 of the API Client contains the method `set_judgment_this_uri` to achieve
this.