Strip left slash from uri when listing objects in an S3 bucket
The background: When moving a "failure" judgment from a failure URI to a
real, NCN-generated URI, the judgment's assets (docx, pdf and images) were
moving to a new prefix in the unpublished assets bucket. But when publishing
the judgment, the assets were not being copied into the published assets bucket.

At first we thought this was due to permissions or encryption on the new
assets. Possinly caused by something to do with permissions differences
between an object created with PutObject, vs an object created with CopyObject.

After much digging, we discovered that the assets were not being published
because the call to `client.list_objects` in `publish_documents()` was always
returning an empty array.

After even more digging, we discovered that `list_objects` doesn't recognise
bucket prefixes that start with a leading slash `/`. Trailing slashes are
fine but leading slashes are not.

For now, trim the leading slash wherever we perform the `list_objects`
action on a bucket.

TODO: somehow remove the leading slash on a URI everywhere.