Add CSS sharing to `fab run`
In https://github.com/nationalarchives/ds-caselaw-editor-ui/pull/355 we added
shared CSS from ds-caselaw-public-ui to this repository via the Dockerfile.

However, some ofthe developers found that they were unable to see the shared
CSS when running the application with `fab run`. `fab build` and `fab start`
build and run the Docker containers using the Dockerfile and retrieve the
latest version of the CSS every time they are run. But `fab run` is more
"lightweight" and does not retrieve the latest CSS every time.

Amend `fab run` to explicityl get the CSS from public-ui, so that developers
using this command to run the application always get the most recent
version of the shared CSS.