Notify Enrichment that a document has been published
When a judgment is edited, a notification is published to the Simple Notification Service ARN at `SNS_TOPIC`. This currently has a MessageAttribute of `update_type` set to "published" or "not published", and a json message of the reference of the document and the current publication state of the document, as specified in the edited form, e.g.

`({"uri_reference": "ewhc/kb/1", "status": "published"})`

Note that, currently, this does NOT mean that the document is freshly published or that it has not been enriched -- these are potential improvements, but will require additional work on modelling the lifecycle of the documents.

This code expects a localstack/AWS SNS (Simple Notification Service) queue to exist; we are using `caselaw-stg-judgment-updated` on staging and will probably use `caselaw-judgment-updated` on production.