Use fields from search results in SearchResult model
This should not be pushed to production before https://github.com/nationalarchives/ds-caselaw-public-access-service/pull/88
is deployed to Marklogic production.

Previously, we were getting a judgement's neutral citation, name, court and date
by making an API call to Marklogic, getting the judgment XML, and parsing it. We
can actually get these elements as part of the search results, and it means we
lose an extra call to the API. As we show at least 10 results per page, this
means the app will load faster, as we lose at least 10 calls per page.