Use SNS Message correctly, by not using MessageStructure
The MessageStructure being set to JSON means that there are multiple
messages, customised for each format. We just want to send one type
of message to all users, which just *happens* to be JSON, so we
shouldn't set MessageStructure. The blank "default" key was just setting
the Message to be blank for all non-recognised users (which was everyone.)

We've also added `uri_reference` to the MessageAttributes. This
was mostly added as it was a way of potentially getting the data
out in some format until I understood what was wrong,
but it'll also let people subscribe to "uri starts with /ewhc/2022"
which is useful.