Bump APIClient version to 5.1.2 and remove all references to XMLTools
XMLTools in the API Client is deprecated. Since we bumped to version 5.1.1 we
have been seeing deprecation warnings in the EUI logs.

In order to remove all references to XMLTools, we had to fix some of the API
methods in the client (see https://github.com/nationalarchives/ds-caselaw-custom-api-client/pull/96)

We now retrieve all metadata for a judgment via API Client calls, not via
extracting the values from the judgment XML using XMLTools.