Fix bug where assignment field empty after editing judgment
This was caused by the fact that the dictionary of user names and ids
was not populated in the handler for the POST request, so the dropdown
was empty on rendering the template. Fixed by factoring the code to generate
this out into a helper method, then calling it from both the POST and GET handler
for judgment edits.