335 - remove unneeded link in breadcrumbs on homepage
The current page in the breadcrumbs should not display as a link, for accessibility and general usability reasons. Previously, the homepage always displayed as a link, whether or not you were currently on that page. This fixes it so that it's only displayed as a link when you are on another page of the site.