Add redirection endpoint for getting signed asset URLs
By setting the value of the `XSLT_IMAGE_LOCATION` environment variable
to this endpoint, when the judgment XML is passed through the XSLT to
generate the HTML it's returned with image locations based around this
endpoint. a GET request to this endpoint will return a HTTP 302 redirect
to the signed S3 URL of the image file in the private bucket.