Update judgment HTML view URL to new structure
This moves the main "show me the content as HTML" view from `/detail`
followed by pulling the judgment URI from the query string, into its own
`/<path>` URL handler. We use `path` because judgment URLs contain
slashes.

As part of this we update the behaviour for 404s; the `Judgment` class
will now try to preload an NCN when it is initialised, which means that
the process of simply initialising a `Judgment` will cause a
`MarklogicResourceNotFoundError` to be thrown if the judgment doesn't
exist. This means we can wrap initialisation in a `try`/`catch` block to
raise 404s before we even try to apply page logic.

This in turn means the "judgment not found" test can actually test this
behaviour, unlike testing Django's own page not found behaviour as it
was before (since `/judgment` didn't exist).