Add toggle switches for HTML/PDF view
These reuse the existing toolbar button components, because there's no
sense reinventing the wheel. It also changes some of the toolbar buttons
to better reflect what they're actually doing in this new context, and
removes "download HTML".

The waffle `{% flag 'foo' %}…{% else %}` behaviour trips up curlylint,
which is why we're catching the flag state in the view and then pushing
it into the render context rather than relying on the tags. If we run
into this more we should consider writing a custom tag so we can query
arbitrary flag states directly and just use them in normal page logic
flows.

Removing `render` from edit_judgment and inlining it with `get` is
because it's not used in any other place, previously it was also used by
`post` before moving to PRG. This makes it more consistent with other
views.