Remove line numbers from location files
Although the file locations can be useful for identifying where a localised string is used, the inclusion of line numbers can lead to very large changesets in the language file as a result of unrelated changes to the code itself.

Remove the line numbers entirely, and update our documentation for the generation command.