Add pytest-socket
This pytest plugin blocks any attempt for tests to contact the network at the socket level.

Our tests should never rely on making an external network call, but it's currently possible (and very easy in functional tests) to accidentally do just that without noticing. pytest-socket prevents any outgoing network calls and immediately raises an exception when they occur, so we can spot the problem and fix it.