Style publish and publish success page
This updates the new publish flow to match Terry's designs. the 'assign' button in the
sidebar and the publish feedback checkboxes are implemented but hidden behind a feature flag for now, pending development of those user stores

We've also created a reusable button-cta css class for consistency, following the pattern introduced with Terry's button-secondary.