Remove doubled 'context' in template parameters.
As well as being confusing this was causing problems for template re-use, where an include was used in two places - one with the extra 'context' variable on the context, and another without. This removes the extra 'context' everywhere it's used for consistency - further refactoring to come to tidy up the views and remove duplication.