Add assignment dropdown to judgment sidebar
This allows editors to reassign judgments on the publish, hold and unhold pages.

It relies on javascript for the behaviour given in the axure prototypes - without JS it will gracefully degrade to a standard form with a submit button.

I've also added test coverage for the assignment button handler logic.