Create '.../move' and '.../overwrite' endpoints
In order to help support press summaries, we want to not automatically
splat existing documents which are sitting at the URI we want to use.
Instead we provide a /move endpoint which will allow documents to be
moved to unoccupied URIs and redirect users to /overwrite if the target
is already taken.

There they will see both documents and metadata for them so they can
decide whether the overwrite is correct or not. Old versions are still
preserved within the document history (see ADR0014, currently draft)