Fix encoding behaviour of email link generators
Previously, this was encoding HTML special characters as the email
template was run through the renderer. Since the emails are sent as
plain text, however, this was leading to things like `&nbsp;` appearing
in the email content.

The only functional change here is a tweak to add `|safe` to the
relevant variables in the template, but to make sure this was tested
there's been a decent chunk of refactoring:

- External link generation is now in its own file
- There are some smoke tests for all the link generators
- The methods are now typed
- Method signatures have been changed so both of the email generation
  methods behave the same way, even though one of them doesn't make use
  of the second variable.
- Email signatures now accept a string, rather than needing to pass in
  the whole request.