Update move/merge language to "change NCN"
This better reflects the path which this form actually presents (ie the document has a valid NCN, and we're going to give it another valid NCN). The case of recovering from a colliding URI is actually slightly different, as the NCN is already valid and we just need to follow this flow to complete the merge.