Support for changes in advance of discrete document types
v12 of the api client introduces some changes to document classes in
readiness to support multiple types. Part of this involves renaming some
aspects so it's clear they relate to the abstract concept of a
`Document`, rather than a specific type. This PR makes the necessary
changes to support those renamings.

This does _not_ move EUI to use discrete document types - it continues
to create `Judgment` objects which behave (almost) identically to
before. A separate PR will move EUI to the point where it receives
arbitrary document classes.