Add a hover color to the tabset, and fix a bug when used twice on same page
Two minor improvements, we now have a hover colour on unselected tabs (and the cursor displays correctly on selected tabs), but also i fixed a slightly puzzling thing where if the tabset is used twice in the same page, the first instance appears to break (showing all tabs as unselected). This has meant a slight change to the interface - if it's used twice on the same page, each tabset must be given a unique `name` property, which is noted in the styleguide.