Remove unused deletion code
The frontend path to reach this was unused, and the code to actually
execute a deletion was badly defined and untested. This removes all
deletion functionality from the EUI in favour of it living in `Document`
objects within the API.

UI to handle deletions may be reimplemented at a later date once we've
ironed out exactly who should have this ability and under what
conditions.