Make use of failed_to_parse over is_failure and is_editable
These state flags did not usefully represent the actual causes for a
document being in its current state.

Documents which have parsed and are only at a failure URI are now
rendered as HTML, since there is no reason not to. They will still not
be publishable, since an incorrect NCN is publish-blocking.