Reinstate editor delete functionality
This is now a link in the 'more drawer', which leads to a confirmation
page. The link will only appear if the document is in a safe to delete
state.

Previously the delete button was only available if the judgment was at a
'failure' URI. This link is currently present in all document contexts,
providing the document is in a safe to delete state. This exposes delete
functionality to all editors for all unpublished documents.