Fix TDR spacing issue in metadata panel
This was a tricky one to track down but not to fix - when the TDR has four characters in the last position, and one of them is em-width (an 'm' or 'w'), at a very specific screensize it pops underneath the status badge and is illegible. The fix is to tweak the breakpoint at which the width of that metadata panel becomes constrained with a max-width, to give it a little more room.

I've also changed the label for the document type from 'Doc type' to 'Type', as discussed on slack - we don't specify 'document' in any other labels here, and it's already unambiguous.