Refactor/improve JudgmentFactory
This change moves `JudgmentFactory` to a more generic `DocumentFactory`,
and implements a `DocumentVersionFactory` which extends the build method
to do version-specific stuff like set version numbers and annotations.

I tried a few approaches to this, including two totally separate
factories, but this seemed the cleanest implementation because of the
way we model versions, namely that they are proper instances of a
`Document` which just happen to have other stuff set as well.

So that we don't get gnarly recursion, `DocumentFactory` now also gains
a `populate_versions` parameter which is set `False` by
`DocumentVersionFactory`, suppressing population of the versions arrays.