Make auto re-enrich aware of parser versions
This also introduces a fix for the case where MarkLogic returns zero
results, in which case instead of returning an array of arrays (with the
first being a header row) it just returns a single array and breaks our
expectations. The new `get_rows_from_result` method in
`judgments.views.reports` will handle this case and will always return
the expected list of results, even if that list is empty.