Watch JS and assets while running the server
When getting set up I was confused about the lack of assets. I then
realised I forgot to build the assets, which is documented a bit
further down.

To make `fab run` more of a one-shot process (and to save us having to
spawn multiple terminal windows), I’ve added a couple of steps that:

- Run `npm install`; and
- Watch the Sass and JS for changes

The `npm` tasks are backgrounded and the log output is in seperate
files. Ideally I’d want to have this logged to stdout along with the
Django logs, but I couldn’t see an easy way to do this in a fabfile.