Always download docx
S3 sets mimetypes if files are uploaded in the console, unlike the ingester. This can prompt
editor's PCs to instead load the documents up in an online version of Office, which fails
for some reason. Force the files to be downloaded when clicking on the links by setting the
`download` attribute.