Remove the EUI trying to move documents on NCN change
This is no longer the desired behaviour; all document URIs are frozen since URL resolution is now handled via the identifiers layer.

For existing documents, this should already be the right NCN (if not we'll resolve this in a future exception report). For new documents, this should be a non-semantic UUID.