Swap NCN for preferred ID in metadata panels
A document's preferred ID may not always be an NCN; change behaviour to reflect this fact, by displaying the preferred identifier and its type (where one exists) and no longer offering the option to manipulate the NCN directly from this form (since that behaviour may not always be valid).

As part of this, we remove a lot of the use of the deprecated `judgment` context, with a view to eventually removing it entirely. This also involves cleaning up the Style Guide and how it gets the `document` context into components. `judgment` is now renamed `mock_document` and is explicitly injected into the components, so there's less magic.