Add display of new TDR fields to document history
To better support editors, add a subset of the new fields from TDR to the document history view.

This is intended as a quick fix to let editors access this new information where it's present; a more holistic approach is still needed to properly integrate these with metadata where necessary.