Add basic telemetry to bulk reparser runs
We want to get better visibility of what's going on when we do bulk reparsing. To that end, we add a new database table (in the EUI) to record run instances and some basic facts about what they did.

As part of this we now also capture more granular information about exact numbers of attempts/skips/failures by capturing exceptions (of known types) during runs and preventing them from crashing the entire process.